proxy: split kube-proxy-ish parts of proxy startup into a separate file
This will make it easier to keep them in sync with kube-proxy.
